com.shr4pnel.clients moved to com.shr4pnel.init
